chapters now work
