+++ app/src/main/java/ckmah/mangoreader/fragment/ChapterListFragment.java
@@ -4,4 +4,6 @@ package ckmah.mangoreader.fragment;
  * Created by shawnxie on 12/24/15.
  */
 public class ChapterListFragment {
+
+
 }
