+++ app/src/main/res/layout/fragment_manga_item_chapters.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="ckmah.mangoreader.fragment.MangaItemChapterFragment">
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/chapter_recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:clipToPadding="false"
+        android:padding="4dp"
+        android:scrollbars="vertical" />
+
+</LinearLayout>
